requirements.txt
h5py==2.10.0
numpy~=1.19.2
six~=1.15.0
mlflow
hydra-core
tensorflow-cpu==2.4.0
keras
